+++ src/main/kotlin/icu/windea/pls/ep/expression/ParadoxLocalisationExpressionSupports.Core.kt
@@ -6,6 +6,7 @@ import com.intellij.openapi.util.*
 import com.intellij.psi.*
 import com.intellij.util.*
 import icu.windea.pls.lang.*
+import icu.windea.pls.lang.codeInsight.completion.*
 import icu.windea.pls.lang.util.*
 import icu.windea.pls.localisation.psi.*
 import icu.windea.pls.lang.expression.complex.*